Created folders for frontend and backend
