--- src/Learn.js
@@ -1,15 +1,0 @@
-import React from 'react'
-import Search from './component/Search.js'
-
-const Learn = () => {
-  return (
-    <>
-        <div>
-          
-          <Search/>
-        </div>
-    </>
-  )
-}
-
-export default Learn
